Agentic_AI_LOGIN2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.5",
    "groq>=1.0.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
